+++ src/snyk/common/languageServer/languageServer.ts
@@ -1,7 +1,6 @@
 import { IAuthenticationService } from '../../base/services/authenticationService';
 import { CLI_INTEGRATION_NAME } from '../../cli/contants/integration';
 import { Configuration, IConfiguration } from '../configuration/configuration';
-import { SNYK_HAS_AUTHENTICATED, SNYK_LANGUAGE_SERVER_NAME } from '../constants/languageServer';
 import { CONFIGURATION_IDENTIFIER } from '../constants/settings';
 import { SNYK_CONTEXT } from '../constants/views';
 import { ErrorHandler } from '../error/errorHandler';
@@ -15,6 +14,9 @@ import { IVSCodeWorkspace } from '../vscode/workspace';
 import { LsExecutable } from './lsExecutable';
 import { LanguageClientMiddleware } from './middleware';
 import { InitializationOptions, LanguageServerSettings } from './settings';
+import { SNYK_CLI_PATH, SNYK_HAS_AUTHENTICATED, SNYK_LANGUAGE_SERVER_NAME } from '../constants/languageServer';
+import { firstValueFrom } from 'rxjs';
+import { DownloadService } from '../services/downloadService';
 
 export interface ILanguageServer {
   start(): Promise<void>;
@@ -33,7 +35,10 @@ export class LanguageServer implements ILanguageServer {
     private window: IVSCodeWindow,
     private authenticationService: IAuthenticationService,
     private readonly logger: ILog,
-  ) {}
+    private downloadService: DownloadService,
+  ) {
+    this.downloadService = downloadService;
+  }
 
   async start(): Promise<void> {
     // TODO remove feature flag when ready
@@ -41,6 +46,9 @@ export class LanguageServer implements ILanguageServer {
       await this.contextService.setContext(SNYK_CONTEXT.PREVIEW_LS_AUTH, false);
       return Promise.resolve(undefined);
     }
+
+    // wait until Snyk LS is downloaded
+    await firstValueFrom(this.downloadService.downloadReady$);
     await this.contextService.setContext(SNYK_CONTEXT.PREVIEW_LS_AUTH, true);
     this.logger.info('Starting Snyk Language Server...');
 
@@ -89,6 +97,12 @@ export class LanguageServer implements ILanguageServer {
 
     this.client.onNotification(SNYK_HAS_AUTHENTICATED, ({ token }: { token: string }) => {
       this.authenticationService.updateToken(token).catch((error: Error) => {
+        ErrorHandler.handle(error, this.logger, error.message);
+      });
+    });
+
+    this.client.onNotification(SNYK_CLI_PATH, ({ cliPath }: { cliPath: string }) => {
+      void this.configuration.setCliPath(cliPath).catch((error: Error) => {
         ErrorHandler.handle(error, this.logger, error.message);
       });
     });