+++ src/snyk/common/languageServer/lsExecutable.ts
@@ -2,25 +2,28 @@ import os from 'os';
 import path from 'path';
 import { Checksum } from '../../cli/checksum';
 import { Platform } from '../platform';
-import { LsSupportedPlatform } from './supportedPlatforms';
+import { LsSupportedPlatform, SupportedLsPlatformsList } from './supportedPlatforms';
+import fs from 'fs/promises';
+import { IConfiguration } from '../configuration/configuration';
 
 export class LsExecutable {
+  private static filenamePrefix = 'snyk-ls';
   public static filenameSuffixes: Record<LsSupportedPlatform, string> = {
-    linux386: 'snyk-ls_linux_386',
-    linuxAmd64: 'snyk-ls_linux_amd64',
-    linuxArm64: 'snyk-ls_linux_arm64',
-    windows386: 'snyk-ls_windows_386.exe',
-    windowsAmd64: 'snyk-ls_windows_amd64.exe',
-    darwinAmd64: 'snyk-ls_darwin_amd64',
-    darwinArm64: 'snyk-ls_darwin_arm64',
+    linux386: 'linux_386',
+    linuxAmd64: 'linux_amd64',
+    linuxArm64: 'linux_arm64',
+    windows386: 'windows_386.exe',
+    windowsAmd64: 'windows_amd64.exe',
+    darwinAmd64: 'darwin_amd64',
+    darwinArm64: 'darwin_arm64',
   };
 
   public static defaultPaths: Record<LsSupportedPlatform, string> = {
     linux386: process.env.XDG_DATA_HOME ?? '/.local/share/',
     linuxAmd64: process.env.XDG_DATA_HOME ?? '/.local/share/',
     linuxArm64: process.env.XDG_DATA_HOME ?? '/.local/share/',
-    windows386: process.env.XDG_DATA_HOME ?? '\\AppData\\Local\\snyk\\',
-    windowsAmd64: process.env.XDG_DATA_HOME ?? '\\AppData\\Local\\snyk\\',
+    windows386: process.env.XDG_DATA_HOME ?? '\\AppData\\Local\\',
+    windowsAmd64: process.env.XDG_DATA_HOME ?? '\\AppData\\Local\\',
     darwinAmd64: process.env.XDG_DATA_HOME ?? '/Library/Application Support/',
     darwinArm64: process.env.XDG_DATA_HOME ?? '/Library/Application Support/',
   };
@@ -28,7 +31,11 @@ export class LsExecutable {
   constructor(public readonly version: string, public readonly checksum: Checksum) {}
 
   static getFilename(platform: LsSupportedPlatform): string {
-    return this.filenameSuffixes[platform];
+    return `${this.filenamePrefix}_${this.filenameSuffixes[platform]}`;
+  }
+
+  static getVersionedFilename(platform: LsSupportedPlatform, version: string) {
+    return `${this.filenamePrefix}_${version}_${this.filenameSuffixes[platform]}`;
   }
 
   static getPath(customPath?: string): string {
@@ -37,11 +44,12 @@ export class LsExecutable {
     }
 
     const platform = this.getCurrentWithArch();
-    const homeDir = Platform.getHomeDir();
-    const lsFilename = this.filenameSuffixes[platform];
-    const defaultPath = this.defaultPaths[platform];
 
-    return path.join(homeDir, defaultPath, lsFilename);
+    const homeDir = Platform.getHomeDir();
+    const lsFilename = this.getFilename(platform);
+    const defaultPath = this.defaultPaths[platform];
+    const lsDir = path.join(homeDir, defaultPath, 'snyk-ls');
+    return path.join(lsDir, lsFilename);
   }
 
   static getCurrentWithArch(): LsSupportedPlatform {
@@ -56,7 +64,17 @@ export class LsExecutable {
     if (opArch === 'ia32') {
       opArch = '386';
     }
+    const supportPlatform = `${opSys}${opArch.charAt(0).toUpperCase()}${opArch.slice(1)}`;
+    if (SupportedLsPlatformsList.find(p => p === supportPlatform) !== undefined) {
+      return supportPlatform as LsSupportedPlatform;
+    }
+    throw new Error(`Unsupported platform: ${supportPlatform}`);
+  }
 
-    return `${opSys}${opArch.charAt(0).toUpperCase()}${opArch.slice(1)}` as LsSupportedPlatform;
+  static exists(configuration: IConfiguration): Promise<boolean> {
+    return fs
+      .access(LsExecutable.getPath(configuration.getSnykLanguageServerPath()))
+      .then(() => true)
+      .catch(() => false);
   }
 }